main.rs: Got basic example of write and read
Got a basic example of reading from one nvim instance and writing to the other
one. To test, run the following in seperate terminals:

nvim --listen 127.0.0.1:6666
nvim --listen 127.0.0.1:6667
cargo run

Type in the instance on port 6667 and you should see the other one fill up
with "hey rust" messages. I'm not handling errors right now, so it breaks
whenever nvim has an issue. In the future we should handle the errors in a
log file and then reattach to the buffer.