main.rs: Refactor parsing
This commit moves all of the parsing logic into it's own function and returns
a struct to represent the nvim buf line events. For now we are copying the
data into the vector for simplicity, but I think we can do better in the future.

It also cleans up a lot of the setup logic and changes many of the unwraps
to handle errors properly.